Move `Histogram<N>` to feature types header
